refactor(tours): add shared tour core helper
